+++ deployment/src/main/resources/dev-templates/embedded.html
@@ -3,7 +3,7 @@
   Logging manager UI</a>
 <br>
 {#if config:property('quarkus.logging-manager.openapi.included') == 'true'}
-<a href="{frameworkRootPath}{config:property('quarkus.swagger-manager.path')}/#/{config:property('quarkus.logging-manager.openapi-tag')}" class="badge badge-light">
+<a href="{frameworkRootPath}{config:property('quarkus.swagger-ui.path')}/#/{config:property('quarkus.logging-manager.openapi-tag')}" class="badge badge-light">
   <i class="fa fa-map-signs fa-fw"></i>
   Spec in Swagger UI</a>
 {/if}
