+++ deployment/src/test/java/io/quarkiverse/loggingmanager/deployment/AbstractConfigurationTest.java
@@ -1,4 +1,4 @@
-package io.quarkiverse.loggingmanager.quarkus.logging.manager.deployment;
+package io.quarkiverse.loggingmanager.deployment;
 
 import static io.restassured.RestAssured.given;
 import static org.hamcrest.MatcherAssert.assertThat;
